==> Issue_Embeddings/notebooks/04b_Inference-Batch.ipynb <==
# Background

This notebook illustrates the use of a utility, `InferenceWrapper.df_to_emb` that can be used to perform inference in bulk on large amounts of data.  A benchmark is provided that compares doing inference one at a time in a serial fashion and demonstrates a **10x speedup in inference time** over the previous method.

# Location of Model Artifacts

### Google Cloud Storage

- **model for inference** (965 MB): `https://storage.googleapis.com/issue_label_bot/model/lang_model/models_22zkdqlr/trained_model_22zkdqlr.hdf`


- **encoder (for fine-tuning w/a classifier)** (965 MB): 
`https://storage.googleapis.com/issue_label_bot/model/lang_model/models_22zkdqlr/trained_model_encoder_22zkdqlr.pth`


- **fastai.databunch** (27.1 GB):
`https://storage.googleapis.com/issue_label_bot/model/lang_model/data_save.hdf`


- **checkpointed model** (2.29 GB): 
`https://storage.googleapis.com/issue_label_bot/model/lang_model/models_22zkdqlr/best_22zkdqlr.pth`

# Load Minimal Model For Inference

from inference import InferenceWrapper, pass_through
from IPython.display import display, Markdown
import pandas as pd
from torch.nn.utils.rnn import pad_sequence
from torch import Tensor, cat, device
from torch.cuda import empty_cache
from typing import List
from tqdm import tqdm
from numpy import concatenate as cat
import torch
import numpy as np

# from fastai.torch_core import defaults
# defaults.device = torch.device('cpu')


#### Create an `InferenceWrapper` object

wrapper = InferenceWrapper(model_path='/ds/Issue-Embeddings/notebooks',
                           model_file_name='trained_model_22zkdqlr.hdf')

#### Download a test dataset

The test dataset has 2,000 GitHub Issues in the below format:

testdf = pd.read_csv(f'https://storage.googleapis.com/issue_label_bot/language_model_data/000000000000.csv.gz').head(8000)

testdf.head(3)



# Perform Batch Inference

Why Batch-Inference?  When there are a large number of issues for which you want to retrieve document embedddings, batch inference on a gpu (should be) significantly faster than on a cpu.

#### Generate Embeddings From Pre-Trained Language Model

See help for `wrapper.df_to_emb`:

help(wrapper.df_to_emb)

### Benchmarking inference time on 8,000 Issues

#### Below is when inference is done using batching (New Method)

%%time
embeddings = wrapper.df_to_emb(testdf)

#### Below is when inference is done one at a time (Old Method)

%%time
# prepare data
test_data = [wrapper.process_dict(x)['text'] for x in testdf.to_dict(orient='rows')]

emb_single = []
for d in tqdm(test_data):
    emb_single.append(wrapper.get_pooled_features(d).detach().cpu().numpy())
    
emb_single_combined = cat(emb_single)

# Notes:

There is a **10x speedup** for inference by chunking the data into batches of similar length (to minimize padding) and passing that through the GPU.

In order to get a further speed improvement we must utilize [pad_packed_sequence](https://pytorch.org/docs/stable/nn.html#torch.nn.utils.rnn.pad_packed_sequence).  We leave this a future exercise to optimize batching more.

### Test

This section tests that the embeddings retrieved from the one-at-a time approach are sufficently close to the embeddings from the batching approach

assert np.allclose(emb_single_combined, embeddings, atol=1e-5)
